+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Add Kubernetes 1.22 and 1.23 to the supported list. Remove tests for Kubernetes 1.19.
 - Upgrade OperatorSDK to v1.9.0.
+- Update CFSSL to 1.5.0-astarte.3.
 
 ### Fixed
 - Force HTTPS for API and dashboard when SSL is enabled.